+++ src/components/shelf/HeroBackground.tsx
@@ -1,9 +1,7 @@
 import { useEffect, useState, useRef } from "react";
 import { getPreferredSteamDocument } from "../../runtime/steamHost";
-import { getRuntimeClassMap } from "../../core/webpackCompat";
+import { logInfo } from "../../runtime/logger";
 
-/** URLs for the hero background — library_hero.jpg is the large landscape art
- *  used by native recents. header.jpg is a smaller fallback. */
 function getHeroUrls(appid: number): string[] {
   return [
     `/customimages/${appid}_hero.png`,
@@ -15,53 +13,66 @@ function getHeroUrls(appid: number): string[] {
   ];
 }
 
+/** Classes discovered from the native recents hero art DOM chain. */
+type NativeHeroClasses = {
+  imgClass: string;
+  zoomContainerClass: string;  // parent of img — carries the 25s zoom animation
+  wrapperClasses: string[];    // intermediate wrappers between outer container and zoom container
+};
+
 export function HeroBackground({ mountEl }: { mountEl: HTMLElement }) {
   const [heroUrl, setHeroUrl] = useState<string | null>(null);
   const [visible, setVisible] = useState(false);
-  const [nativeImgClass, setNativeImgClass] = useState('');
-  const [nativeContainerClass, setNativeContainerClass] = useState('');
+  const [nativeClasses, setNativeClasses] = useState<NativeHeroClasses | null>(null);
   const imgRef = useRef<HTMLImageElement>(null);
   const fallbackIdx = useRef(0);
   const currentAppid = useRef(0);
   const allUrls = useRef<string[]>([]);
 
-  // Discover native hero img classes from the recents section for theme compatibility
+  // Discover native hero classes from the recents section
   useEffect(() => {
-    const discoverClasses = () => {
+    const discover = () => {
       const doc = getPreferredSteamDocument();
       if (!doc) return;
+      // The recents section is the sibling before our mount
       const prev = mountEl.previousElementSibling as HTMLElement | null;
       if (!prev) return;
-      // Find the large hero image in native recents
       const imgs = Array.from(prev.querySelectorAll('img'));
       for (const img of imgs) {
         const r = img.getBoundingClientRect();
         if (r.width > 400 && r.height > 200) {
-          setNativeImgClass(Array.from(img.classList).join(' '));
-          if (img.parentElement) {
-            setNativeContainerClass(Array.from(img.parentElement.classList).join(' '));
+          const imgClass = Array.from(img.classList).join(' ');
+          // Walk up: parent is the zoom container (25s animation), then intermediate wrappers
+          const zoomContainer = img.parentElement;
+          if (!zoomContainer) continue;
+          const zoomContainerClass = Array.from(zoomContainer.classList).join(' ');
+          // Collect intermediate wrappers up to the top-level Focusable
+          const wrapperClasses: string[] = [];
+          let el = zoomContainer.parentElement;
+          for (let i = 0; i < 4 && el && el !== prev; i++) {
+            if (el.classList.contains('Focusable') || el.classList.contains('Panel')) break;
+            wrapperClasses.push(Array.from(el.classList).join(' '));
+            el = el.parentElement;
           }
+          setNativeClasses({ imgClass, zoomContainerClass, wrapperClasses });
+          logInfo("HOME", "HeroBackground: native classes discovered", { imgClass: imgClass.substring(0, 40), zoomContainerClass: zoomContainerClass.substring(0, 40) });
           return;
         }
       }
     };
-    discoverClasses();
-    const t = setTimeout(discoverClasses, 2000);
+    discover();
+    const t = setTimeout(discover, 2000);
     return () => clearTimeout(t);
   }, [mountEl]);
 
   useEffect(() => {
     const updateHero = () => {
-      // Find focused card in the first shelf only
       const firstShelf = mountEl.querySelector('.ds-shelf');
       if (!firstShelf) { setVisible(false); return; }
-
       const focused = firstShelf.querySelector('.ds-card.gpfocus, .ds-card:focus') as HTMLElement | null;
       if (!focused) { setVisible(false); return; }
-
       const appid = Number(focused.getAttribute('data-appid') ?? 0);
       if (appid <= 0) { setVisible(false); return; }
-
       if (appid !== currentAppid.current) {
         currentAppid.current = appid;
         const urls = getHeroUrls(appid);
@@ -77,15 +88,11 @@ export function HeroBackground({ mountEl }: { mountEl: HTMLElement }) {
     const observer = new MutationObserver(updateHero);
     observer.observe(mountEl, { subtree: true, attributes: true, attributeFilter: ['class'] });
     mountEl.addEventListener('focusin', updateHero);
-
     const onFocusOut = (e: FocusEvent) => {
       const related = e.relatedTarget as HTMLElement | null;
-      if (!related || !mountEl.contains(related)) {
-        setVisible(false);
-      }
+      if (!related || !mountEl.contains(related)) setVisible(false);
     };
     mountEl.addEventListener('focusout', onFocusOut);
-
     return () => {
       observer.disconnect();
       mountEl.removeEventListener('focusin', updateHero);
@@ -104,9 +111,44 @@ export function HeroBackground({ mountEl }: { mountEl: HTMLElement }) {
 
   if (!heroUrl) return null;
 
-  // The hero extends from above the mount (negative top) to fill the area
-  // that native recents would occupy. position:absolute relative to the
-  // deck-shelves-root which has position:relative.
+  // Replicate the native DOM structure so theme CSS rules match:
+  // outer (clip) > wrappers > zoomContainer (25s zoom anim) > img (grayscale, fade-in anim)
+  const wrappers = nativeClasses?.wrapperClasses ?? [];
+
+  let inner = (
+    <div
+      className={nativeClasses?.zoomContainerClass || undefined}
+      style={{ position: "absolute", inset: 0, overflow: "visible" }}
+    >
+      <img
+        ref={imgRef}
+        className={nativeClasses?.imgClass || undefined}
+        src={heroUrl}
+        onError={onImgError}
+        style={{
+          width: "100%",
+          height: "100%",
+          objectFit: "cover",
+          objectPosition: "50% 50%",
+          display: "block",
+        }}
+        loading="eager"
+      />
+    </div>
+  );
+
+  // Wrap in intermediate containers (native has 2-3 wrappers with absolute positioning)
+  for (let i = wrappers.length - 1; i >= 0; i--) {
+    inner = (
+      <div
+        className={wrappers[i] || undefined}
+        style={{ position: "absolute", inset: 0, overflow: "visible" }}
+      >
+        {inner}
+      </div>
+    );
+  }
+
   return (
     <div
       className="ds-hero-background"
@@ -123,31 +165,7 @@ export function HeroBackground({ mountEl }: { mountEl: HTMLElement }) {
         transition: "opacity 0.5s cubic-bezier(0.17, 0.45, 0.14, 0.83)",
       }}
     >
-      <div
-        className={nativeContainerClass || undefined}
-        style={{
-          position: "absolute",
-          inset: 0,
-          overflow: "hidden",
-        }}
-      >
-        <img
-          ref={imgRef}
-          className={nativeImgClass || undefined}
-          src={heroUrl}
-          onError={onImgError}
-          style={{
-            width: "100%",
-            height: "100%",
-            objectFit: "cover",
-            objectPosition: "50% 50%",
-            display: "block",
-            transition: "opacity 0.5s cubic-bezier(0.17, 0.45, 0.14, 0.83)",
-          }}
-          loading="eager"
-        />
-      </div>
-      {/* Gradient fade at the bottom to blend with content below */}
+      {inner}
       <div style={{
         position: "absolute",
         bottom: 0,
